Rename shapes_1d -> signals
