+++ tools/train/train_signals.py
@@ -2,7 +2,7 @@ import torch
 
 from generation.config import SIGNALS_TRAINING_CONFIG as CONFIG
 from generation.dataset.signals_dataset import SignalsDataset
-from generation.nets.shapes_1d import Generator, Discriminator
+from generation.nets.signals import Generator, Discriminator
 from generation.training.wgan_trainer import WganTrainer
 from generation.training.utils import set_seed
